tools/infrastructure-dependency-assessment/apps/reporter/requirements.txt
matplotlib==3.10.8
Pillow==12.1.1
pytest==9.0.2
python_docx==1.2.0
qc==0.1
